chore: reuse index page
