--- packages/gasket-plugin-cypress/generator/test/e2e/index.cy.js
@@ -1,8 +1,0 @@
-/* eslint-disable */
-
-describe('IndexPage', () => {
-  it('renders page', () => {
-    cy.visit('/');
-    cy.get('h1').contains('Welcome to Gasket!');
-  });
-});
