+++ packages/gasket-plugin-nextjs/generator/app/app-router/app/page.js
@@ -1,9 +1,4 @@
-/* eslint-disable no-unused-vars */
-import React from 'react';
-import GasketEmblem from '@gasket/assets/react/gasket-emblem';
-
-const pageStyle = { textAlign: 'center' };
-const logoStyle = { width: '250px', height: '250px' };
+import IndexPage from '../components/index-page';
 
 export const metadata = {
   title: 'Home',
@@ -11,13 +6,4 @@ export const metadata = {
   charset: 'UTF-8'
 };
 
-export default function Page() {
-  return (
-    <div style={ pageStyle }>
-      <GasketEmblem style={ logoStyle }/>
-      <h1>Welcome to Gasket!</h1>
-      <p>To get started, edit <code>pages/index.js</code> and save to reload.</p>
-      <p><a href='https://gasket.dev'>Learn Gasket</a></p>
-    </div>
-  );
-}
+export default IndexPage
