+++ packages/gasket-plugin-nextjs/generator/app/shared/README.md
