+++ packages/gasket-plugin-nextjs/package.json
@@ -107,6 +107,7 @@
   "eslintIgnore": [
     "generator/redux/redux/store.js",
     "generator/app/pages-router/pages/_app.js",
+    "generator/app/shared/components/index-page.js",
     "*.d.ts"
   ],
   "babel": {
